# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/dummyLogin": {
      "post": {
        "summary": "Dummy Login",
        "operationId": "dummy_login_dummyLogin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DummyLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешная авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pvz": {
      "post": {
        "summary": "Create Pvz",
        "operationId": "create_pvz_pvz_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PVZCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ПВЗ создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PVZ"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Pvz",
        "operationId": "list_pvz_pvz_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начальная дата диапазона",
              "title": "Start Date"
            },
            "description": "Начальная дата диапазона"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конечная дата диапазона",
              "title": "End Date"
            },
            "description": "Конечная дата диапазона"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "Количество элементов на странице",
              "default": 10,
              "title": "Limit"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PVZNested"
                  },
                  "title": "Response List Pvz Pvz Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receptions": {
      "post": {
        "summary": "Create Reception",
        "operationId": "create_reception_receptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Приемка создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reception"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос или есть незакрытая приемка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/products": {
      "post": {
        "summary": "Add Product",
        "operationId": "add_product_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Товар добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос или нет активной приемки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/pvz/{pvz_id}/delete_last_product": {
      "post": {
        "summary": "Delete Last Product",
        "operationId": "delete_last_product_pvz__pvz_id__delete_last_product_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pvz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pvz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар удален",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Неверный запрос, нет активной приемки или нет товаров для удаления",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pvz/{pvz_id}/close_last_reception": {
      "post": {
        "summary": "Close Last Reception",
        "operationId": "close_last_reception_pvz__pvz_id__close_last_reception_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pvz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pvz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Приемка закрыта",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reception"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос или приемка уже закрыта",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register User",
        "operationId": "register_user_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login User",
        "operationId": "login_user_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешная авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DummyLoginRequest": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "employee",
              "moderator"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "DummyLoginRequest"
      },
      "Error": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Error"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PVZ": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "registration_date": {
            "type": "string",
            "format": "date-time",
            "title": "Registration Date"
          },
          "city": {
            "type": "string",
            "enum": [
              "Москва",
              "Санкт-Петербург",
              "Казань"
            ],
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "id",
          "registration_date",
          "city"
        ],
        "title": "PVZ"
      },
      "PVZCreate": {
        "properties": {
          "city": {
            "type": "string",
            "enum": [
              "Москва",
              "Санкт-Петербург",
              "Казань"
            ],
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "city"
        ],
        "title": "PVZCreate"
      },
      "PVZNested": {
        "properties": {
          "pvz": {
            "$ref": "#/components/schemas/PVZ"
          },
          "receptions": {
            "items": {
              "$ref": "#/components/schemas/ReceptionNested"
            },
            "type": "array",
            "title": "Receptions"
          }
        },
        "type": "object",
        "required": [
          "pvz",
          "receptions"
        ],
        "title": "PVZNested"
      },
...
      "ProductCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "электроника",
              "одежда",
              "обувь"
            ],
            "title": "Type"
          },
          "pvz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pvz Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "pvz_id"
        ],
        "title": "ProductCreate"
      },
      "Reception": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "pvz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pvz Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "close"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date_time",
          "pvz_id",
          "status"
        ],
        "title": "Reception"
      },
      "ReceptionCreate": {
        "properties": {
          "pvz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pvz Id"
          }
        },
        "type": "object",
        "required": [
          "pvz_id"
        ],
        "title": "ReceptionCreate"
      },
...
      "Token": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "Token"
      },
...
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}